daemon: move comment `jsonrpc_` methods to an individual module
